pyproject.toml
[tool.poetry]
name = "webtool"
version = "0.0.2"
description = "FastAPI/Starlette library for authentication, throttling, caching."
authors = ["IIH <ds5105119@gmail.com>"]
readme = "README.md"
packages = [{include = "webtool"}]

[tool.poetry.dependencies]
python = "^3.12"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
redis = {extras = ["hiredis"], version = "^5.1.1"}
msgspec = "^0.18.6"
blake3 = "^0.4.1"
argon2-cffi = "^23.1.0"


[tool.poetry.group.dev.dependencies]
fastapi = {extras = ["standard"], version = "^0.115.2"}
pytest-asyncio = "^0.24.0"
requests = "^2.32.3"
coverage = "^7.6.4"
ruff = "^0.7.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.pytest.ini_options]
asyncio_default_fixture_loop_scope = "session"